Create JSON file of main Breakdance DIV
If you have breakdance you might be able to copy this json then paste it directly to your editor.  It should create a div wrapper with all of the listboxes, search button and post loop builder etc.

The only thing missing will be the global block that the post loop builder uses to display content.